Do not use UDF types as return values

+++ examples/udf_1dim.py
@@ -7,7 +7,7 @@ from time import time
 import numpy as np
 
 import iarray as ia
-from iarray.udf import jit, Array, float64, int64
+from iarray.udf import jit, Array, float64
 
 
 # Number of iterations per benchmark
@@ -21,7 +21,7 @@ ia.set_config_defaults(favor=ia.Favor.SPEED, dtype=dtype)
 
 
 @jit
-def f(out: Array(float64, 1), x: Array(float64, 1)) -> int64:
+def f(out: Array(float64, 1), x: Array(float64, 1)) -> int:
     n = out.shape[0]
     for i in range(n):
         out[i] = (math.sin(x[i]) - 1.35) * (x[i] - 4.45) * (x[i] - 8.5)
